Convert MCP schemas to strict where possible
## Summary:
Towards #404. I made this configurable because it's not clear this is always a good thing to do. I also made it default to False because I'm not sure if this could cause errors.

If it works out well, we can switch the default in the future as a small breaking changes

## Test Plan:
Unit tests